Add run and pace conversion

+++ boot/src/main/kotlin/org/freekode/tp2intervals/domain/activity/Activity.kt
@@ -10,7 +10,7 @@ data class Activity(
     val type: TrainingType,
     val title: String,
     val duration: Duration,
-    val load: Long?,
+    val load: Int?,
     val resource: String?
 ) {
     private fun compareDuration(otherDuration: Duration): Boolean {
